pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=2.0",
    "numpy>=2.2.6",
    "paho-mqtt>=2.1.0",
    "requests>=2.28",
]
